Немножко разделяем вьюхи
