+++ backend/api/views/__init__.py
@@ -1,0 +1,10 @@
+from api.views.ingredient import IngredientViewSet
+from api.views.recipe import RecipeRedirectView, RecipeViewSet
+from api.views.tag import TagViewSet
+
+__all__ = [
+    'IngredientViewSet',
+    'RecipeRedirectView',
+    'RecipeViewSet',
+    'TagViewSet'
+]
