+++ backend/api/views/ingredient.py
@@ -1,0 +1,12 @@
+from rest_framework import viewsets
+from rest_framework.permissions import AllowAny
+
+from api.models import Ingredient
+from api.serializers import IngredientSerializer
+
+
+class IngredientViewSet(viewsets.ReadOnlyModelViewSet):
+    queryset = Ingredient.objects.all()
+    serializer_class = IngredientSerializer
+    permission_classes = [AllowAny]
+    pagination_class = None
