+++ backend/api/views/tag.py
@@ -1,0 +1,12 @@
+from rest_framework import viewsets
+from rest_framework.permissions import AllowAny
+
+from api.models import Tag
+from api.serializers import TagSerializer
+
+
+class TagViewSet(viewsets.ReadOnlyModelViewSet):
+    queryset = Tag.objects.all()
+    serializer_class = TagSerializer
+    permission_classes = [AllowAny]
+    pagination_class = None
