+++ backend/users/views/__init__.py
@@ -1,0 +1,5 @@
+from users.views.user import UserViewSet
+
+__all__ = [
+    'UserViewSet'
+]
